Extract controllers from app.js into the controllers/ folder.
