+++ app.js
@@ -5,8 +5,7 @@ const views = require('koa-views');
 const serve = require('koa-static');
 const config = require('./config');
 const db = require('./db');
-// This library help us to interact with post requests
-const parse = require('co-body');
+const controllers = require('./controllers');
 
 const TEMPORAL_DIR = path.join(process.cwd(), '.tmp');
 
@@ -40,97 +39,9 @@ function createApplication(controllers) {
 
 // create and run app
 createApplication({
-  homeController: homeController(),
-  talkController: talkController(),
-  userController: userController(),
-  donationController: donationController(),
-  eventController: eventController()
+  homeController: controllers.homeController(db),
+  talkController: controllers.talkController(db),
+  userController: controllers.userController(db),
+  donationController: controllers.donationController(db),
+  eventController: controllers.eventController(db)
 });
-
-const homeController = function() {
-  return {
-    renderHome: function* () {
-      yield this.render('index', {});
-    },
-    renderAboutUs: function* () {
-      yield this.render('about_us', {});
-    },
-    renderContact: function* () {
-      yield this.render('contact', {});
-    },
-    pageNotFound: function* () {
-      if( this.status === 404){
-        yield this.render('404');
-      }
-    }
-  };
-};
-
-const talkController = function () {
-  return {
-    renderTalkSubmissionForm: function* () {
-      const userGroups = yield db.usergroups.find();
-      const context = {usergroups: userGroups};
-
-      yield this.render('submit_talk', context);
-    },
-    submitTalk: function* () {
-      // TODO: Only allow a set of specific fields.
-      const rawBody = yield parse(this);
-      const belongsToUserGroup = rawBody.usergroups && rawBody.usergroups.length;
-
-      // TODO: Render an error message if talk doesn't belong to an user group.
-      if (belongsToUserGroup) {
-        yield db.talks.create(rawBody);
-      }
-
-      return this.redirect('/speakers');
-    }
-  };
-};
-
-const userController = function () {
-  return {
-    usergroup: function* (){
-      var group =  new db.usergroups({
-        name: 'Python Dominicana',
-        facebook_url: 'https://www.facebook.com/groups/pythondo/',
-        website: 'www.python.com.do',
-        logo: 'https://scontent-mia1-1.xx.fbcdn.net/hphotos-ash2/v/t1.0-9/10624973_10204014303853300_9205003069849326284_n.jpg?oh=7bcdfcda40ccb89341d73ebfe9bcdbb6&oe=570E52D0',
-        created: new Date('2013-09-25T19:00:00.001Z'),
-        description: 'Python Dominicana is an amazing place, is the place where Programmers, engineers an all kind of people interested in the Python Computer Language get together to talk, learn and share about all the stuff that can be done with it, from web applications to robots, from distributed systems to desktop apps.'
-      });
-      group.save(function(err){
-        if(err) throw err;
-      });
-
-      this.body = 'OK';
-    },
-    users: function* () {
-      db.users.insert({
-        first_name: 'Vivian',
-        last_name: 'Guillen',
-        email: 'vivian@codetiger.co',
-        admin: true
-      }, function(err){
-        if(err) throw err;
-      });
-    }
-  };
-};
-
-const donationController = function () {
-  return {
-    renderDonations: function* (){
-      yield this.render('donations', {});
-    }
-  };
-};
-
-const eventController = function () {
-  return {
-    renderEvents: function* (){
-      yield this.render('events', {});
-    }
-  };
-};
