+++ controllers/donationController.js
@@ -1,0 +1,9 @@
+function donationController() {
+  return {
+    renderDonations: function* (){
+      yield this.render('donations', {});
+    }
+  };
+}
+
+module.exports = donationController;
