+++ controllers/eventController.js
@@ -1,0 +1,9 @@
+function eventController() {
+  return {
+    renderEvents: function* (){
+      yield this.render('events', {});
+    }
+  };
+}
+
+module.exports = eventController;
