+++ controllers/index.js
@@ -1,0 +1,13 @@
+const donationController = require('./donationController');
+const eventController = require('./eventController');
+const homeController = require('./homeController');
+const talkController = require('./talkController');
+const userController = require('./userController');
+
+module.exports = {
+  donationController,
+  eventController,
+  homeController,
+  talkController,
+  userController
+};
